+++ template/src/components/DevPanel/DevPanel.tsx
@@ -1,8 +1,10 @@
-import React from "react"
-import { DevToolsPanel } from "@tesler-ui/core"
+import React from 'react'
+import { DevToolsPanel } from '@tesler-ui/core'
+import { useSelector } from 'react-redux';
+import { AppState } from '../../interfaces/storeSlices'
 
 const DevPanel: React.FunctionComponent = () => {
-    const showCondition = process.env.NODE_ENV === 'development'
+    const showCondition = useSelector((state: AppState) => state.session.devPanelEnabled)
     return (
         <>
             {showCondition && (