+++ template/src/reducers/session.ts
@@ -2,20 +2,23 @@ import { actionTypes, AnyAction } from '../interfaces/actions'
 import {AppState} from '../interfaces/storeSlices'
 import {Session} from '@tesler-ui/core/interfaces/session'
 
+export type CustomSession = Session & { logout: boolean }
+
 /**
  * Your initial state for this slice
  */
-export const initialState: Session = {
+export const initialState: CustomSession = {
     active: false,
     screens: [],
-    loginSpin: false
+    loginSpin: false,
+    logout: false
 }
 
 export default function sessionReducer(
-    state: Session = initialState,
+    state: CustomSession = initialState,
     action: AnyAction,
     store?: Readonly<AppState>
-): Session {
+): CustomSession {
     switch (action.type) {
         /**
          * Your reducers for this slice
@@ -24,7 +27,15 @@ export default function sessionReducer(
             return {
                 ...state,
                 loginSpin: false,
-                active: false
+                active: false,
+                logout: true
+            }
+        }
+        case actionTypes.loginDone: {
+            return {
+                ...state,
+                active: true,
+                logout: false
             }
         }
         default: